CSS tweaks so cancel can be clicked also in BUY/SELL ORDERS sections
The cancel button was already visible for every order in the "BUY
ORDERS" and "SELL ORDERS" section, but due to an unfortunate CSS
definition it was present in a zero-width column, which made it
completely unclickable. I have redefined column widths for 5-column
tables (a 5-column table is actually a 3-column table, as the first one
and the last one is always empty).

The actual visual layout hasn't changed much. I couldn't find any other
5-column table which could've been broken by this change.