Disabled accepting orders on closed markets
When a market was closed, you could still accept orders in "BUY ORDERS"
and "SELL ORDERS" sections by clicking on them. It was confusing and
due to it clicking the cancel button was difficult. I blocked it so these
orders can now only be listed or cancelled when the market is closed.