Order cancellation messages fixed in case of a closed market
When a market is closed, any order can be cancelled. It means that we can
cancel an order which hasn't been submitted by us. In that case, the message
in the cancellation popup should be rephrased to say that the tokens will
be returned to the owner of the order, not to the person who is cancelling it.

Also, the "If someone (partially) fills this order before you cancel it,
your order may not or only be partially cancelled." mesage should be visible
only when the market is still open. In case of a closed market, this situation
will never happen.