Hiding dust orders
Hiding orders below a certain threshold in the "BUY ORDERS" and "SELL ORDERS"
sections. The "MY ORDERS" section still shows all orders submitted by us, even
if they are below that threshold.

The threshold is defined per currency (although I presume it's only W-ETH
currently) and is hardcoded for now. I've set it to 0.001 W-ETH.

In order to implement collection filtering efficiently I had to store 'buyHowMuch'
and 'sellHowMuch' fields as numbers as well ie. by creating two additional fields
in the Offer document. It doesn't give us perfect precision, but it's for
filtering only so should be ok. The same trick is already being used for
sorting as well.