Incremental lookback for historical trades fetching
In order to optimize OasisDEX Pro page load times, I have changed the
Trade event fetching mechanism so we do fetch only the last weeks Trades.

In order to find the block range of Trades to be fetched,
I did implement a simple lookback algorithm. Please be aware, that the
'startBlockNumber' could be a bit earlier than the beginning (midnight)
of the first day we are interested in, so I did apply some additional
filtering in the actual trade fetching in order to ignore the too-early
ones.