Gas consumption estimation in order placement modal
As the new SimpleMarket contract becomes more unpredictable in terms of
gas consumption (it depends on the number of orders matched, sorting
etc.), I have added gas consumption estimation in the order placement
modal. The estimated value (with some markup on top of it) is used as
the actual gas value during order placement as well.