Temporary dapple build until we can generate Meteor templates for dapp builds
I have switched to dapp-built versions of maker-otc and token-wrapper.
Had to create a temporary structure to generate Meteor templates using
dapple though (and keep the contract addresses somewhere as well), as we
still cannot do it with dapp (and it's likely dapp will never be able to
do it and we will have to come up with some other solution).

I have also linked to automated-matching enabled version of the
maker-otc contract by @r001.