Historical trade scanning modified to use LogTake events instead of Trade events
Switched to using LogTake events. Since we now have timestamps in
LogTake events, which we didn't have in Trade events, it is not needed
anymore to query individual blocks in order to fetch the timestamps.
This should make trade history loading faster, proportionally
to the number of trades made on the contract.